PLANET-8160 Refactor Spreadsheet block to utilize block.json
This is to take advantage of Gutenberg's way of defining blocks
