--- assets/src/blocks/Spreadsheet/SpreadsheetEditorScript.js
@@ -1,3 +1,0 @@
-import {registerSpreadsheetBlock} from './SpreadsheetBlock';
-
-registerSpreadsheetBlock();
