+++ assets/src/scss/blocks.scss
@@ -9,7 +9,6 @@
 @import "blocks/ActionsList/ActionsListStyle";
 @import "blocks/PostsList/PostsListStyle";
 @import "blocks/SocialMedia/SocialMediaStyle";
-@import "blocks/Spreadsheet";
 @import "blocks/TableOfContents/TableOfContentsStyle";
 @import "blocks/TakeActionBoxout/TakeActionBoxoutStyle";
 @import "blocks/Timeline/TimelineStyle";
