+++ src/Blocks/Rest.php
@@ -9,6 +9,10 @@ use WP_REST_Server;
  */
 class Rest
 {
+    // Variables for the Spreadsheet block functions.
+    private const SPREADSHEET_MAX_ROWS = 10000;
+    private const SPREADSHEET_CACHE_LIFETIME = 30;
+
     public function load(): void
     {
         add_action('rest_api_init', function (): void {
@@ -28,6 +32,38 @@ class Rest
                         'methods' => WP_REST_Server::READABLE,
                         'callback' => static function ($fields) {
                             return rest_ensure_response(self::get_happypoint_data($fields));
+                        },
+                    ],
+                ]
+            );
+
+            /**
+             * Endpoint to retrieve a Spreadsheet data and cache it.
+             */
+            register_rest_route(
+                BaseBlock::REST_NAMESPACE,
+                '/get-spreadsheet-data',
+                [
+                    [
+                        'permission_callback' => static function () {
+                            return true;
+                        },
+                        'methods' => WP_REST_Server::READABLE,
+                        'callback' => static function () {
+                            $sheet_id = filter_input(
+                                INPUT_GET,
+                                'sheet_id',
+                                FILTER_VALIDATE_REGEXP,
+                                [
+                                    'options' => [
+                                        'regexp' => '/[\w\d\-]+/',
+                                    ],
+                                ]
+                            );
+
+                            $sheet_data = self::get_spreadsheet($sheet_id, false);
+
+                            return rest_ensure_response($sheet_data);
                         },
                     ],
                 ]
@@ -60,5 +96,72 @@ class Rest
         $data['default_embed_code'] = $options['happy_point_embed_code'] ?? '';
 
         return $data;
+    }
+
+    /**
+     * Fetch a Google sheet by its ID. Used by the Spreadsheet block.
+     *
+     * @param string|null $sheet_id The ID of the Google sheet.
+     * @param bool        $skip_cache Should the sheet be fetched from cache.
+     * @return array|null The sheet or null if nothing was found.
+     */
+    public static function get_spreadsheet(?string $sheet_id, bool $skip_cache): ?array
+    {
+        if (! $sheet_id) {
+            return null;
+        }
+
+        $cache_key = "spreadsheet_{$sheet_id}";
+
+        if (! $skip_cache) {
+            $from_cache = wp_cache_get($cache_key);
+
+            if (false !== $from_cache) {
+                return $from_cache;
+            }
+        }
+
+        $url = "https://docs.google.com/spreadsheets/d/e/{$sheet_id}/pub?output=csv";
+
+        $headers = get_headers($url);
+
+        // Handle 500, 404 errors.
+        if (! $headers || strpos($headers[0], '500') || strpos($headers[0], '404')) {
+            return null;
+        }
+
+		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_fopen
+        $handle = fopen($url, 'rb');
+
+        if (false === $handle) {
+            return null;
+        }
+
+        $rows = [];
+        while (
+            // phpcs:ignore Squiz.PHP.DisallowSizeFunctionsInLoops,WordPress.CodeAnalysis.AssignmentInCondition
+            ( $data = fgetcsv($handle, 1000, ',') ) !== false
+        ) {
+            if (count($rows) > self::SPREADSHEET_MAX_ROWS) {
+                break;
+            }
+
+            $rows[] = $data;
+        }
+		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_fclose
+        fclose($handle);
+
+        if (0 === count($rows)) {
+            $sheet = null;
+        } else {
+            $sheet = [
+                'header' => $rows[0],
+                'rows' => array_slice($rows, 1),
+            ];
+        }
+
+        wp_cache_add($cache_key, $sheet, null, self::SPREADSHEET_CACHE_LIFETIME);
+
+        return $sheet;
     }
 }
